+++ tests/testthat/test-nipals.R
@@ -2,13 +2,66 @@
 
 # see setup-expectations.R for expect_aligned
 
-test_that("Fitted values", {
-  # Published examples of NIPALS with missing values are 
-  # extremely rare.  Here is an example of estimating the missing values
-  # using XLSTAT (details on methodology are limited)
+# Note, I looked at the python repository:
+# https://github.com/fredrikw/python-nipals/blob/master/tests/test_nipals.py
+# I did not understand the tests and so did not incorporate the yarn
+# data or oliveoil data.
+
+## ---------------------------------------------------------------------------
+
+# Create data for tests
+
+B_full <- matrix(c(50, 67, 90, 98, 120,
+                   55, 71, 93, 102, 129,
+                   65, 76, 95, 105, 134,
+                   50, 80, 102, 130, 138,
+                   60, 82, 97, 135, 151,
+                   65, 89, 106, 137, 153,
+                   75, 95, 117, 133, 155), ncol=5, byrow=TRUE)
+rownames(B_full) <- c("G1","G2","G3","G4","G5","G6","G7")
+colnames(B_full) <- c("E1","E2","E3","E4","E5")
+
+# Same data with 2 NAs
+B_miss <- matrix(c(NA, 67, 90, 98, 120,
+                   NA, 71, 93, 102, 129,
+                   65, 76, 95, 105, 134,
+                   50, 80, 102, 130, 138,
+                   60, 82, 97, 135, 151,
+                   65, 89, 106, 137, 153,
+                   75, 95, 117, 133, 155), ncol=5, byrow=TRUE)
+rownames(B_miss) <- c("G1","G2","G3","G4","G5","G6","G7")
+colnames(B_miss) <- c("E1","E2","E3","E4","E5")
+
+## ---------------------------------------------------------------------------
+
+test_that("For complete-data matrix, fitted values match original data", {
+
+  m1 <- nipals(B_full, center=TRUE, scale=TRUE, fitted=TRUE, tol=1e-8)
+  expect_equal( round( m1$fitted, 3), B_full )
+
+  m2 <- nipals(B_full, center=TRUE, scale=FALSE, tol=1e-8, fitted=TRUE)
+  expect_equal( round( m2$fitted, 3), B_full )
+  
+  m3 <- nipals(B_full, center=FALSE, scale=FALSE, tol=1e-8, fitted=TRUE)
+  expect_equal( round( m3$fitted, 3), B_full )
+  
+  m4 <- nipals(B_full, center=FALSE, scale=TRUE, tol=1e-8, fitted=TRUE)
+  expect_equal( round( m4$fitted, 3), B_full )
+    
+})
+
+test_that("Estimate fitted values for NAs", {
+  # Published examples of NIPALS with missing values are almost nonexistant.
+  # Here is an example with XLSTAT (details on methodology are limited)
+  # 
+  # Circa 2017 XLSTAT estimated the missing values below as
   # https://help.xlstat.com/customer/en/portal/articles/2062415-missing-data-imputation-using-nipals-in-excel?b_id=9283
-  # Circa 2017 XLSTAT estimated the missing values below as 1365.2, 88.6 etc
-  # In 2019 XLSTAT generates different values.
+  #   1365.236, 88.600, 175.798, 1051.698, 432.470, 168.554
+  #
+  # In 2019, XLSTAT estimates
+  # https://help.xlstat.com/6497-missing-data-imputation-using-nipals-excel
+  #   1398.829, 88.151, 178.313, 1060.534, 444.254, 169.686
+  #
   auto <- data.frame(capacity = c(1396, 1721, 1580, 1769, 2068, 1769L),
                      power = c(90, 92, 83, 90, 88, 90L),
                      speed = c(174, 180, 170, 180, 180, 182L),
@@ -32,63 +85,51 @@ test_that("Fitted values", {
 })
 
 test_that("Code coverage of nipals function arguments", {
-  B <- matrix(c(50, 67, 90, 98, 120,
-                55, 71, 93, 102, 129,
-                65, 76, 95, 105, 134,
-                50, 80, 102, 130, 138,
-                60, 82, 97, 135, 151,
-                65, 89, 106, 137, 153,
-                75, 95, 117, 133, 155), ncol=5, byrow=TRUE)
-  rownames(B) <- c("G1","G2","G3","G4","G5","G6","G7")
-  colnames(B) <- c("E1","E2","E3","E4","E5")
   
   library(nipals)
-  Bnarow = B
+  Bnarow = B_full
   Bnarow[1,] = NA
   expect_error(nipals(Bnarow))
-  Bnacol = B
+  Bnacol = B_full
   Bnacol[,1] = NA
   expect_error(nipals(Bnacol))
   
-  B2 = B
-  B2[1,1] = B2[2,1] = NA
-  
   # ncomp
-  m1 = nipals(B2, ncomp=1)
+  m1 = nipals(B_miss, ncomp=1)
   
   # center/scale
-  m1 = nipals(B2, center=FALSE, scale=FALSE)
-  m1 = nipals(B2, center=TRUE, scale=FALSE)
-  m1 = nipals(B2, center=TRUE, scale=TRUE)
+  m1 = nipals(B_miss, center=FALSE, scale=FALSE)
+  m1 = nipals(B_miss, center=TRUE, scale=FALSE)
+  m1 = nipals(B_miss, center=TRUE, scale=TRUE)
   
   # maxiter
-  expect_warning(nipals(B, maxiter=2))
+  expect_warning(nipals(B_full, maxiter=2))
 
   # tol
-  m1 = nipals(B2, tol=1e-1, verbose=TRUE)
-  m1 = nipals(B2, tol=1e-10, verbose=TRUE)
+  m1 = nipals(B_miss, tol=1e-1, verbose=TRUE)
+  m1 = nipals(B_miss, tol=1e-10, verbose=TRUE)
 
   # startcol
-  m1 = nipals(B, startcol=0, verbose=TRUE)
-  m1 = nipals(B, startcol=5, verbose=TRUE)
+  m1 = nipals(B_full, startcol=0, verbose=TRUE)
+  m1 = nipals(B_full, startcol=5, verbose=TRUE)
 
   # fitted
-  expect_null(nipals(B)$fitted)
-  expect_null(nipals(B, fitted=FALSE)$fitted)
-  expect_false(is.null(nipals(B, fitted=TRUE)$fitted))
+  expect_null(nipals(B_full)$fitted)
+  expect_null(nipals(B_full, fitted=FALSE)$fitted)
+  expect_false(is.null(nipals(B_full, fitted=TRUE)$fitted))
 
   # force.na
-  m1 = nipals(B, force.na=FALSE)
-  m1 = nipals(B, force.na=TRUE)
+  m1 = nipals(B_full, force.na=FALSE)
+  m1 = nipals(B_full, force.na=TRUE)
   
   # gramschmidt
-  m1 = nipals(B2, gramschmidt=FALSE) # default
+  m1 = nipals(B_miss, gramschmidt=FALSE) # default
   round(crossprod(m1$loadings), 3)  # 1 on diagonal, but not identity
-  m2 = nipals(B2, gramschmidt=TRUE)
+  m2 = nipals(B_miss, gramschmidt=TRUE)
   round(crossprod(m2$loadings), 3)  # should be identity 5x5
   
   # verbose
-  m1 = nipals(B, verbose=TRUE)
+  m1 = nipals(B_full, verbose=TRUE)
   
 })
 
@@ -115,7 +156,7 @@ test_that("Start column function", {
   
 })
 
-test_that("Predictions from model", {
+test_that("Calculate predictions from model", {
   
   # choose 75% of rows for training sample
   set.seed(42)
@@ -151,26 +192,15 @@ test_that("Predictions from model", {
   
 })
 
-# ----------------------------------------------------------------------------
-## expect_allclose <- function(A, B, rtol=1e-5, atol=1e-8){
-##   # https://docs.scipy.org/doc/numpy/reference/generated/numpy.allclose.html
-##   # a, b, rtol=1e-05, atol=1e-08, equal_nan=False
-##   # Returns True if two arrays are element-wise equal within a tolerance.
-##   # The tolerance values are positive, typically very small numbers.
-##   # The relative difference (rtol * abs(b)) and the absolute difference atol
-##   # are added together to compare against the absolute difference
-##   # between a and b.
-
-##   # If either array contains one or more NaNs, False is returned. Infs are treated as equal if they are in the same place and of the same sign in both arrays.
-  
-##   # If the following equation is element-wise True, then allclose returns True.
-##   # absolute(a - b) <= (atol + rtol * absolute(b))
-##   # The above equation is not symmetric in a and b, so that allclose(a, b)
-##   # might be different from allclose(b, a) in some rare cases.
-
-##   diff <- abs(A-B) < (atol + rtol * abs(B))
-##   print(diff)
-##   expect_true(all(diff))
-## }
-## #expect_allclose(m1s$u, m1n$scores, rtol=1e-5, atol=1e-8)
-
+test_that("avg_angular_distance", {
+  #  https://math.stackexchange.com/questions/2113634/
+  rot1 <- matrix(c(-0.956395958, -0.292073218, 0.000084963,
+                   0.292073230, -0.956395931, 0.000227268,
+                   0.000014880, 0.000242173, 0.999999971),
+                 ncol=3, byrow=TRUE)
+  rot2 <- matrix(c(-0.956227882, -0.292623029, -0.000021887,
+                   0.292623030, -0.956227882, -0.000024473,
+                   -0.000013768, -0.000029806, 0.999999999),
+                 ncol=3, byrow=TRUE)
+  expect_equal(avg_angular_distance(rot1, rot2), .0004950387)
+})